Renames middleware
