+++ lib/middleware/messages/member/catch-all.js
@@ -2,7 +2,7 @@
 
 const helpers = require('../../../helpers');
 
-module.exports = function continueConversation() {
+module.exports = function catchAllReply() {
   return (req, res) => {
     if (!helpers.request.hasCampaign(req)) {
       return helpers.replies.noCampaign(req, res);
